+++ npm/webpack-dev-server/src/makeDefaultWebpackConfig.ts
