+++ npm/webpack-dev-server/src/makeWebpackConfig.ts
@@ -2,7 +2,7 @@ import { debug as debugFn } from 'debug'
 import * as path from 'path'
 import { merge } from 'webpack-merge'
 import type { Configuration, EntryObject } from 'webpack'
-import { makeCypressWebpackConfig } from './cypressWebpackConfig'
+import { makeCypressWebpackConfig } from './makeDefaultWebpackConfig'
 import type { CreateFinalWebpackConfig } from './createWebpackDevServer'
 import { configFiles } from './constants'
 import { dynamicAbsoluteImport, dynamicImport } from './dynamic-import'
